# [FSConnector] improve performance with asyncio and direct read to memory object