--- a/lmcache/v1/storage_backend/connector/fs_connector.py
+++ b/lmcache/v1/storage_backend/connector/fs_connector.py
@@ -17,6 +17,10 @@ from pathlib import Path
 from typing import List, Optional, no_type_check
 import asyncio
 
+# Third Party
+import aiofiles
+import aiofiles.os
+
 # First Party
 from lmcache.logging import init_logger
 from lmcache.utils import CacheEngineKey
@@ -67,40 +71,39 @@ class FSConnector(RemoteConnector):
     async def exists(self, key: CacheEngineKey) -> bool:
         """Check if key exists in file system"""
         file_path = self._get_file_path(key)
-        return file_path.exists()
+        return await aiofiles.os.path.exists(file_path)
 
     async def get(self, key: CacheEngineKey) -> Optional[MemoryObj]:
         """Get data from file system"""
         file_path = self._get_file_path(key)
-        if not file_path.exists():
-            return None
 
         try:
-            # Read file content
-            with open(file_path, "rb") as f:
-                data = f.read()
+            async with aiofiles.open(file_path, "rb") as f:
+                # Read metadata buffer first to get shape, dtype, fmt
+                # to be able to allocate memory object for the data and read into it
+                md_buffer = bytearray(METADATA_BYTES_LEN)
+                num_read = await f.readinto(md_buffer)
+                if num_read != len(md_buffer):
+                    raise RuntimeError(
+                        f"Partial read meta {len(md_buffer)} got {num_read}"
+                    )
 
-            # Split metadata and actual data
-            metadata = RemoteMetadata.deserialize(memoryview(data[:METADATA_BYTES_LEN]))
-            kv_bytes = data[METADATA_BYTES_LEN : METADATA_BYTES_LEN + metadata.length]
+                # Deserialize metadata and allocate memory
+                metadata = RemoteMetadata.deserialize(md_buffer)
+                memory_obj = self.local_cpu_backend.allocate(
+                    metadata.shape, metadata.dtype, metadata.fmt
+                )
+                if memory_obj is None:
+                    logger.debug("Memory allocation failed during async disk load.")
+                    return None
 
-            # Allocate memory and copy data
-            memory_obj = self.local_cpu_backend.allocate(
-                metadata.shape,
-                metadata.dtype,
-                metadata.fmt,
-            )
-            if memory_obj is None:
-                logger.warning("Failed to allocate memory during file read")
-                return None
-
-            if isinstance(memory_obj.byte_array, memoryview):
-                view = memory_obj.byte_array
-                if view.format == "<B":
-                    view = view.cast("B")
-            else:
-                view = memoryview(memory_obj.byte_array)
-            view[: metadata.length] = kv_bytes
+                # Read the actual data into allocated memory
+                buffer = memory_obj.byte_array
+                num_read = await f.readinto(buffer)
+                if num_read != len(buffer):
+                    raise RuntimeError(
+                        f"Partial read data {len(buffer)} got {num_read}"
+                    )
 
             return memory_obj
 
@@ -115,25 +118,26 @@ class FSConnector(RemoteConnector):
 
         try:
             # Prepare metadata
+            buffer = memory_obj.byte_array
             metadata = RemoteMetadata(
-                len(memory_obj.byte_array),
+                len(buffer),
                 memory_obj.get_shape(),
                 memory_obj.get_dtype(),
                 memory_obj.get_memory_format(),
             )
 
             # Write to file (metadata + data)
-            with open(temp_path, "wb") as f:
-                f.write(metadata.serialize())
-                f.write(memory_obj.byte_array)
+            async with aiofiles.open(temp_path, "wb") as f:
+                await f.write(metadata.serialize())
+                await f.write(buffer)
 
             # Atomically rename temp file to final destination
-            temp_path.replace(final_path)
+            await aiofiles.os.replace(temp_path, final_path)
 
         except Exception as e:
             logger.error(f"Failed to write file {final_path}: {str(e)}")
-            if temp_path.exists():
-                temp_path.unlink()  # Remove corrupted file
+            if await aiofiles.os.path.exists(temp_path):
+                await aiofiles.os.unlink(temp_path)  # Remove corrupted file
             raise
 
     @no_type_check