@guymguym 2025-05-31
@maobaolong I separated the printing changes and will submit those in a separate PR.
Please review again, thanks.